requirements.txt
-i https://pypi.org/simple
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
exceptiongroup==1.1.2 ; python_version < '3.11'
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pluggy==1.2.0 ; python_version >= '3.7'
py==1.11.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0 ; python_version >= '3.7'
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1 ; python_version < '3.11'
urllib3==2.0.4 ; python_version >= '3.7'
